+++ src/main/java/io/soluble/pjb/bridge/MethodCache.java
@@ -31,7 +31,7 @@ import java.util.Map;
  * synchronization, so use this class per thread or per request
  * only.
  */
-final class MethodCache {
+public final class MethodCache {
     Map map;
     static final Entry noCache = new NoCache();